+++ fr/main/model/commanders/FakeCommander.java
@@ -5,7 +5,7 @@ import fr.main.model.Player;
 public class FakeCommander extends Commander {
 
   public FakeCommander (Player player) {
-    super(player, new Power(1), new Power(2));
+    super(player, new Power(1000), new Power(2000));
   }
 
 }